fix profile dialog orientation crash
