+++ app/src/main/java/org/oppia/app/profileprogress/ProfileProgressFragment.kt
@@ -9,7 +9,8 @@ import org.oppia.app.fragment.InjectableFragment
 import javax.inject.Inject
 
 /** Fragment that displays profile progress in the app. */
-class ProfileProgressFragment : InjectableFragment() {
+class ProfileProgressFragment : InjectableFragment(),
+  ProfilePictureClickListener {
   companion object {
     internal const val PROFILE_PROGRESS_FRAGMENT_PROFILE_ID_KEY =
       "ProfileProgressFragment.internal_profile_id"
@@ -37,9 +38,14 @@ class ProfileProgressFragment : InjectableFragment() {
     container: ViewGroup?,
     ProfileProgressFragment: Bundle?
   ): View? {
+    this.retainInstance = true
     val args =
       checkNotNull(arguments) { "Expected arguments to be passed to ProfileProgressFragment" }
     val internalProfileId = args.getInt(PROFILE_PROGRESS_FRAGMENT_PROFILE_ID_KEY, -1)
     return profileProgressFragmentPresenter.handleCreateView(inflater, container, internalProfileId)
+  }
+
+  override fun onProfilePictureClicked() {
+    profileProgressFragmentPresenter.showPictureEditDialog()
   }
 }
