+++ app/src/main/java/org/oppia/app/profileprogress/ProfileProgressFragmentPresenter.kt
@@ -14,6 +14,8 @@ import org.oppia.app.recyclerview.BindableAdapter
 import org.oppia.app.viewmodel.ViewModelProvider
 import javax.inject.Inject
 
+private const val TAG_PROFILE_PICTURE_EDIT_DIALOG = "PROFILE_PICTURE_EDIT_DIALOG"
+
 /** The presenter for [ProfileProgressFragment]. */
 @FragmentScope
 class ProfileProgressFragmentPresenter @Inject constructor(
@@ -83,5 +85,15 @@ class ProfileProgressFragmentPresenter @Inject constructor(
   private enum class ViewType {
     VIEW_TYPE_HEADER,
     VIEW_TYPE_RECENTLY_PLAYED_STORY
+  }
+
+  fun showPictureEditDialog() {
+    val previousFragment =
+      activity.supportFragmentManager.findFragmentByTag(TAG_PROFILE_PICTURE_EDIT_DIALOG)
+    if (previousFragment != null) {
+      activity.supportFragmentManager.beginTransaction().remove(previousFragment).commitNow()
+    }
+    val dialogFragment = ProfilePictureEditDialogFragment.newInstance()
+    dialogFragment.showNow(activity.supportFragmentManager, TAG_PROFILE_PICTURE_EDIT_DIALOG)
   }
 }
